requirements.txt
astunparse==1.6.3
filelock==3.18.0
fsspec==2025.5.1
Jinja2==3.1.6
jittor==1.3.9.14
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-ml-py==12.575.51
nvitop==1.5.1
opencv-python==4.6.0.66
pandas==2.3.0
pillow==11.2.1
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
